+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Higher-level wrappers
+  - backend init/free
+  - model load/free
+  - context init/free/reset
+  - tokenize/detokenize
 
 ## [0.0.1] - 2023-10-16
 ### Added